Move SubnetAllocator into a subdirectory
A copy of the code is used by ovn-kubernetes too, so make it more
shareable again.
