+++ pkg/network/master/util/subnet_allocator_test.go
@@ -1,4 +1,4 @@
-package master
+package util
 
 import (
 	"fmt"
